utils/dashboard/dashboard/requirements.txt
dash
dash-bootstrap-components
flask_cache
pandas
numpy